requirements.txt
pandas>=2.0
scikit-learn>=1.4
joblib>=1.3
pydantic>=2.0
pydantic-ai>=0.0.20
python-dotenv>=1.0
sentence-transformers>=3.0
